chore: restore zero-lint baseline and add lint to npm run build
Fix all 79 lint errors on main:
- Mechanical: prefer-const, hoist functions before useStableEvent,
  replace ternary expression statements with if/else
- no-explicit-any in 14 test files: replace as any casts with
  Partial<ProjectStore>, proper Project model types, and
  FeatureWithClonedDef
- max-lines in featureSlice.ts (1271→under 1200): extract
  buildShapeFeature() helper, consolidating 9 duplicated shape
  constructors into thin wrappers
- max-lines in SketchCanvas.tsx (4019→under 3800): extract
  drawStlTopViewImage to stlTopViewRenderer.ts, local types to
  SketchCanvas.types.ts, and triggerDimensionEdit to its own module

Make lint part of the build gate:
- package.json: npm run build now runs lint first, so CI fails on
  future regressions
- AGENTS.md: updated Build & Verify to mention lint

Closes #223
