+++ src/components/canvas/SketchCanvas.tsx
@@ -15,14 +15,11 @@
  */
 
 import { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react'
-import type { ToolpathResult } from '../../engine/toolpaths/types'
 import { ToolpathVisibilityPanel } from '../ToolpathVisibilityPanel'
-import type { ToolpathVisibility } from '../toolpathVisibility'
-import type { SnapMode, SnapSettings } from '../../sketch/snapping'
-import type { OpenProfileEndpoint, SketchControlRef, SketchEditTool } from '../../store/types'
+import type { OpenProfileEndpoint, SketchControlRef } from '../../store/types'
 import { useProjectStore } from '../../store/projectStore'
 import { previewOffsetFeatures } from '../../store/helpers/derivedFeatures'
-import { chamferDistanceFromPoint, chamferFeatureFromDistance, chamferFeatureFromPoint, filletFeatureFromPoint, filletFeatureFromRadius, filletRadiusFromPoint, mirrorFeatureFromReference, resizeBackdropFromReference, resizeFeatureFromReference, rotateBackdropFromReference, rotateFeatureFromReference } from '../../store/helpers/referenceTransforms'
+import { chamferFeatureFromDistance, chamferFeatureFromPoint, filletFeatureFromPoint, filletFeatureFromRadius, mirrorFeatureFromReference, resizeBackdropFromReference, resizeFeatureFromReference, rotateBackdropFromReference, rotateFeatureFromReference } from '../../store/helpers/referenceTransforms'
 import {
   buildPendingDraftProfile,
   buildPendingProfile,
@@ -42,8 +39,6 @@ import {
 import {
   computeDimensionEditPreviewPoint,
   computeMoveDistancePreviewPoint,
-  computeRotateDegreesFromPreview,
-  computeScaleFactorFromPreview,
 } from './manualEntry'
 import type { OperationDimEdit } from './manualEntry'
 import { useDimensionEditWorkflow } from './useDimensionEditWorkflow'
@@ -87,7 +82,23 @@ import {
   distance2,
   segmentHitTest,
 } from './hitTest'
+import { drawStlTopViewImage } from './stlTopViewRenderer'
+import { triggerDimensionEdit as triggerDimensionEditFn } from './triggerDimensionEdit'
 import { resolveProfileSegments } from '../../store/helpers/resolveProfileSegments'
+import {
+  HANDLE_HIT_RADIUS,
+  MIN_SKETCH_ZOOM,
+  NODE_HIT_RADIUS,
+  OPEN_ENDPOINT_JOIN_HIT_RADIUS,
+  POLYGON_CLOSE_RADIUS,
+  type OpenEndpointHit,
+  type PendingPreviewPoint,
+  type PendingSketchFillet,
+  type SegmentHit,
+  type SketchCanvasHandle,
+  type SketchCanvasProps,
+  type SketchEditPreviewPoint,
+} from './SketchCanvas.types'
 import { segmentIntersections, type ResolvedSeg } from '../../store/helpers/segmentIntersection'
 import { arcControlPoint, anchorPointForIndex, traceProfilePath } from './profilePrimitives'
 import {
@@ -109,7 +120,7 @@ import {
   profileVertices,
   rectProfile,
 } from '../../types/project'
-import type { Clamp, OperationKind, Point, SketchFeature, Tab } from '../../types/project'
+import type { Clamp, Point, SketchFeature, Tab } from '../../types/project'
 import { compatibleFeatureIdsForOperation } from '../cam/operationValidity'
 import { formatLength, parseLengthInput } from '../../utils/units'
 import { useAxisLock, lockModeGuideColor } from '../../sketch/useAxisLock'
@@ -126,131 +137,7 @@ import {
   type FeatureClipboardPayload,
 } from '../../platform/featureClipboard'
 
-const NODE_HIT_RADIUS = 9
-const HANDLE_HIT_RADIUS = 7
-const POLYGON_CLOSE_RADIUS = 12
-const OPEN_ENDPOINT_JOIN_HIT_RADIUS = 14
-const MIN_SKETCH_ZOOM = 0.02
-
-interface PendingPreviewPoint {
-  point: Point
-  session: number
-}
-
-interface SketchEditPreviewPoint {
-  point: Point
-  mode: SketchEditTool
-}
-
-interface PendingSketchFillet {
-  anchorIndex: number
-  corner: Point
-}
-
-interface OpenEndpointHit {
-  featureId: string
-  endpoint: OpenProfileEndpoint
-  anchor: Point
-}
-
-interface SegmentHit {
-  segmentIndex: number
-  point: Point
-}
-
-export interface SketchCanvasHandle {
-  zoomToModel: () => void
-}
-
-interface SketchCanvasProps {
-  onFeatureContextMenu?: (featureId: string, x: number, y: number) => void
-  onTabContextMenu?: (tabId: string, x: number, y: number) => void
-  onClampContextMenu?: (clampId: string, x: number, y: number) => void
-  toolpaths?: ToolpathResult[]
-  selectedOperationId?: string | null
-  collidingClampIds?: string[]
-  snapSettings: SnapSettings
-  zoomWindowActive?: boolean
-  onZoomWindowComplete?: () => void
-  onActiveSnapModeChange?: (mode: SnapMode | null) => void
-  depthLegendCollapsed?: boolean
-  onToggleDepthLegend?: () => void
-  toolpathVisibility?: ToolpathVisibility
-  onToolpathVisibilityChange?: (visibility: ToolpathVisibility) => void
-  /**
-   * A1.3: when an operation kind is armed/hovered in the CAM "Add operation"
-   * menu, the canvas highlights features that operation could act on and dims
-   * the rest. Null when nothing is armed.
-   */
-  operationHighlightKind?: OperationKind | null
-}
-
-function drawStlTopViewImage(
-  ctx: CanvasRenderingContext2D,
-  feature: SketchFeature,
-  image: HTMLImageElement,
-  vt: ViewTransform,
-  selected: boolean,
-  hovered: boolean,
-  editing: boolean,
-): void {
-  if (!image.complete || image.naturalWidth <= 0 || image.naturalHeight <= 0) return
-
-  const verts = profileVertices(feature.sketch.profile)
-  if (verts.length < 3) return
-
-  const angle = ((feature.sketch.orientationAngle ?? 0) * Math.PI) / 180
-  const ux = Math.cos(angle)
-  const uy = Math.sin(angle)
-  const vx = -Math.sin(angle)
-  const vy = Math.cos(angle)
-
-  let minU = Infinity, maxU = -Infinity
-  let minV = Infinity, maxV = -Infinity
-  for (const point of verts) {
-    const projectedU = point.x * ux + point.y * uy
-    const projectedV = point.x * vx + point.y * vy
-    if (projectedU < minU) minU = projectedU
-    if (projectedU > maxU) maxU = projectedU
-    if (projectedV < minV) minV = projectedV
-    if (projectedV > maxV) maxV = projectedV
-  }
-
-  const width = maxU - minU
-  const height = maxV - minV
-  if (!(width > 1e-9) || !(height > 1e-9)) return
-
-  const centerU = minU + width / 2
-  const centerV = minV + height / 2
-  const center = worldToCanvas({
-    x: ux * centerU + vx * centerV,
-    y: uy * centerU + vy * centerV,
-  }, vt)
-  const drawW = width * vt.scale
-  const drawH = height * vt.scale
-
-  ctx.save()
-  traceProfilePath(ctx, feature.sketch.profile, vt)
-  ctx.clip('evenodd')
-  ctx.translate(center.cx, center.cy)
-  ctx.rotate(angle)
-  ctx.globalAlpha = selected || hovered || editing ? 0.72 : 0.86
-  ctx.drawImage(image, -drawW / 2, -drawH / 2, drawW, drawH)
-  ctx.restore()
-
-  ctx.save()
-  traceProfilePath(ctx, feature.sketch.profile, vt)
-  ctx.strokeStyle = selected
-    ? '#efbc7a'
-    : hovered
-      ? '#d2a064'
-      : editing
-        ? '#f7cd87'
-        : '#bcc8d4'
-  ctx.lineWidth = selected || editing ? 2.5 : 1.8
-  ctx.stroke()
-  ctx.restore()
-}
+export type { SketchCanvasHandle }
 
 export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(function SketchCanvas(
   {
@@ -2018,7 +1905,6 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
   //      gains nothing, so it stays — and ESLint reports it as an "unused directive"
   //      warning, which is the (intentional) cost of the bail.
   // Full analysis + reproducible reorder script: planning/archive/LINT_HOOK_TYPING_DEBT_Plan.md.
-  // eslint-disable-next-line react-hooks/exhaustive-deps -- file-wide React-Compiler bail; masks forward-refs + ref-mirror writes + hot-path render-time ref reads (see comment above; planning/archive/LINT_HOOK_TYPING_DEBT_Plan.md)
   }, [projectKey])
 
   useEffect(() => {
@@ -2496,7 +2382,7 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
                               },
                             }
                           }
-                          let previewHits = segmentIntersections(
+                          const previewHits = segmentIntersections(
                             extension,
                             tgtSeg,
                             { rayA: true },
@@ -2527,7 +2413,7 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
                               : subjSeg.a1 - TWO_PI,
                           }
                         }
-                        let previewHits = segmentIntersections(
+                        const previewHits = segmentIntersections(
                           extension,
                           tgtSeg,
                           { rayA: true },
@@ -2649,171 +2535,32 @@ export const SketchCanvas = forwardRef<SketchCanvasHandle, SketchCanvasProps>(fu
   // Called by the "Type" button in banners — opens dimension input without
   // the Tab toggle-close behaviour. Keyboard Tab keeps its existing logic.
   function triggerDimensionEdit() {
-    const project = projectRef.current
-    const pendingAdd = pendingAddRef.current
-    const pendingMove = pendingMoveRef.current
-    const pendingTransform = pendingTransformRef.current
-    const pendingOffset = pendingOffsetRef.current
-    const selection = selectionRef.current
-    const units = project.meta.units
-
-    if (pendingAdd) {
-      if (
-        (pendingAdd.shape === 'rect' || pendingAdd.shape === 'circle' || pendingAdd.shape === 'ellipse' || pendingAdd.shape === 'tab' || pendingAdd.shape === 'clamp' || pendingAdd.shape === 'roundrect' || pendingAdd.shape === 'chamferrect')
-        && pendingAdd.anchor
-      ) {
-        const previewPoint = pendingPreviewPointRef.current?.point ?? pendingAdd.anchor
-        if (pendingAdd.shape === 'circle') {
-          const r = Math.hypot(previewPoint.x - pendingAdd.anchor.x, previewPoint.y - pendingAdd.anchor.y)
-          dimEdit.setDimensionEdit({ shape: 'circle', anchor: pendingAdd.anchor, signX: 1, signY: 1, activeField: 'radius', width: '', height: '', radius: formatLength(r, units), length: '', angle: '' })
-        } else {
-          const w = Math.abs(previewPoint.x - pendingAdd.anchor.x)
-          const h = Math.abs(previewPoint.y - pendingAdd.anchor.y)
-          const dimShape = (pendingAdd.shape === 'roundrect' || pendingAdd.shape === 'chamferrect') ? 'rect' : pendingAdd.shape
-          dimEdit.setDimensionEdit({ shape: dimShape, anchor: pendingAdd.anchor, signX: previewPoint.x >= pendingAdd.anchor.x ? 1 : -1, signY: previewPoint.y >= pendingAdd.anchor.y ? 1 : -1, activeField: 'width', width: formatLength(w, units), height: formatLength(h, units), radius: '', length: '', angle: '' })
-        }
-        return
-      }
-      if ((pendingAdd.shape === 'polygon' || pendingAdd.shape === 'spline') && pendingAdd.points.length >= 1) {
-        const fromPoint = pendingAdd.points[pendingAdd.points.length - 1]
-        const previewPoint = pendingPreviewPointRef.current?.point ?? fromPoint
-        const dx = previewPoint.x - fromPoint.x
-        const dy = previewPoint.y - fromPoint.y
-        dimEdit.setDimensionEdit({ shape: pendingAdd.shape, anchor: fromPoint, signX: 1, signY: 1, activeField: 'length', width: '', height: '', radius: '', length: formatLength(Math.hypot(dx, dy), units), angle: (Math.atan2(dy, dx) * (180 / Math.PI)).toFixed(2).replace(/\.?0+$/, '') })
-        return
-      }
-      if (pendingAdd.shape === 'composite' && pendingAdd.start && !pendingAdd.closed) {
-        if (pendingAdd.currentMode === 'arc' && pendingAdd.pendingArcEnd) {
-          const arcStart = pendingAdd.lastPoint ?? pendingAdd.start
-          const arcEnd = pendingAdd.pendingArcEnd
-          const previewPoint = pendingPreviewPointRef.current?.point
-          const halfChord = Math.hypot(arcEnd.x - arcStart.x, arcEnd.y - arcStart.y) / 2
-          let r = halfChord
-          if (previewPoint) {
-            const midX = (arcStart.x + arcEnd.x) / 2
-            const midY = (arcStart.y + arcEnd.y) / 2
-            const chordDx = arcEnd.x - arcStart.x
-            const chordDy = arcEnd.y - arcStart.y
-            const chordLen = Math.hypot(chordDx, chordDy)
-            if (chordLen > 1e-9) {
-              const perpX = -chordDy / chordLen
-              const perpY = chordDx / chordLen
-              const bulge = (previewPoint.x - midX) * perpX + (previewPoint.y - midY) * perpY
-              r = Math.max(halfChord, Math.abs(bulge) > 1e-9
-                ? (halfChord * halfChord + bulge * bulge) / (2 * Math.abs(bulge))
-                : halfChord)
-            }
-          }
-          const side = previewPoint ? (() => {
-            const cross = (arcEnd.x - arcStart.x) * (previewPoint.y - arcStart.y)
-              - (arcEnd.y - arcStart.y) * (previewPoint.x - arcStart.x)
-            return cross >= 0 ? 1 : -1
-          })() : 1
-          dimEdit.setDimensionEdit({ shape: 'composite', anchor: arcStart, arcStart, arcEnd, arcClockwise: side < 0, signX: 1, signY: 1, activeField: 'radius', width: '', height: '', radius: formatLength(r, units), length: '', angle: '' })
-          return
-        }
-        const fromPoint = pendingAdd.lastPoint ?? pendingAdd.start
-        const previewPoint = pendingPreviewPointRef.current?.point ?? fromPoint
-        const dx = previewPoint.x - fromPoint.x
-        const dy = previewPoint.y - fromPoint.y
-        const len = Math.hypot(dx, dy)
-        const angleDeg = (Math.atan2(dy, dx) * (180 / Math.PI)).toFixed(2).replace(/\.?0+$/, '')
-        const defaultRadius = pendingAdd.currentMode === 'arc' ? formatLength(len > 1e-9 ? len : 0.5, units) : ''
-        dimEdit.setDimensionEdit({ shape: 'composite', anchor: fromPoint, signX: 1, signY: 1, activeField: 'length', width: '', height: '', radius: defaultRadius, length: formatLength(len, units), angle: angleDeg })
-        return
-      }
-      if (pendingAdd.shape === 'slot' && pendingAdd.points.length === 1) {
-        const p1 = pendingAdd.points[0]
-        const previewPoint = pendingPreviewPointRef.current?.point
-        const dx = previewPoint ? previewPoint.x - p1.x : 0
-        const dy = previewPoint ? previewPoint.y - p1.y : 0
-        const len = Math.hypot(dx, dy)
-        const defaultLen = len > 1e-10 ? len : (units === 'mm' ? 20 : 1)
-        const angleDeg = len > 1e-10 ? (Math.atan2(dy, dx) * (180 / Math.PI)).toFixed(2).replace(/\.?0+$/, '') : '0'
-        dimEdit.setDimensionEdit({ shape: 'slot', anchor: p1, arcStart: p1, signX: 1, signY: 1, activeField: 'length', length: formatLength(defaultLen, units), angle: angleDeg, radius: formatLength(units === 'mm' ? 6 : 0.25, units), width: '', height: '' })
-        return
-      }
-      if (pendingAdd.shape === 'slot' && pendingAdd.points.length >= 2) {
-        const p1 = pendingAdd.points[0]
-        const p2 = pendingAdd.points[1]
-        const previewPoint = pendingPreviewPointRef.current?.point
-        const axisX = p2.x - p1.x
-        const axisY = p2.y - p1.y
-        const axisLen = Math.hypot(axisX, axisY)
-        const currentWidth = previewPoint && axisLen > 1e-10
-          ? Math.max(2 * Math.abs((previewPoint.x - p1.x) * axisY - (previewPoint.y - p1.y) * axisX) / axisLen, 0.001)
-          : (units === 'mm' ? 6 : 0.25)
-        dimEdit.setDimensionEdit({ shape: 'slot', anchor: p1, arcStart: p1, arcEnd: p2, signX: 1, signY: 1, activeField: 'width', width: formatLength(currentWidth, units), height: '', radius: '', length: '', angle: '' })
-        return
-      }
-      if (pendingAdd.shape === 'ngon' && pendingAdd.anchor) {
-        const previewPoint = pendingPreviewPointRef.current?.point ?? pendingAdd.anchor
-        const r = Math.hypot(previewPoint.x - pendingAdd.anchor.x, previewPoint.y - pendingAdd.anchor.y)
-        const angleDeg = (Math.atan2(previewPoint.y - pendingAdd.anchor.y, previewPoint.x - pendingAdd.anchor.x) * (180 / Math.PI)).toFixed(2).replace(/\.?0+$/, '')
-        dimEdit.setDimensionEdit({ shape: 'ngon', anchor: pendingAdd.anchor, signX: 1, signY: 1, activeField: 'radius', width: '', height: '', radius: formatLength(r, units), length: '', angle: angleDeg })
-        return
-      }
-    }
-
-    if (pendingMove?.fromPoint && !pendingMove.toPoint) {
-      const previewPoint = pendingMovePreviewPointRef.current?.point ?? pendingMove.fromPoint
-      move.beginMoveDistanceEntry(previewPoint)
-      return
-    }
-
-    if (pendingTransform?.mode === 'resize' && pendingTransform.referenceStart && pendingTransform.referenceEnd) {
-      let factor = '1'
-      const previewPoint = pendingTransformPreviewPointRef.current?.point
-      if (previewPoint) factor = computeScaleFactorFromPreview(pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-      setOperationDimEdit({ kind: 'scale', factor })
-      return
-    }
-
-    if (pendingTransform?.mode === 'rotate' && pendingTransform.referenceStart && pendingTransform.referenceEnd) {
-      let angle = '0'
-      const previewPoint = pendingTransformPreviewPointRef.current?.point
-      if (previewPoint) angle = computeRotateDegreesFromPreview(pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-      setOperationDimEdit({ kind: 'rotate', angle })
-      return
-    }
-
-    if (pendingOffset) {
-      let distance = '0'
-      const rawOffsetPoint = pendingOffsetRawPreviewPointRef.current?.point
-      const snappedOffsetPoint = pendingOffsetPreviewPointRef.current?.point
-      if (rawOffsetPoint && snappedOffsetPoint) {
-        const canvas = canvasRef.current
-        if (canvas) {
-          const vt = computeViewTransform(project.stock, canvas.width, canvas.height, viewStateRef.current)
-          const sourceFeatures = pendingOffset.entityIds
-            .map((id) => project.features.find((f) => f.id === id) ?? null)
-            .filter((f): f is SketchFeature => f !== null)
-            .filter((f) => f.sketch.profile.closed)
-          const previewInput = resolveOffsetPreview(sourceFeatures, rawOffsetPoint, snappedOffsetPoint, snap.activeSnapRef.current?.mode ?? null, vt)
-          if (previewInput) distance = formatLength(previewInput.signedDistance, units)
-        }
-      }
-      setOperationDimEdit({ kind: 'offset', distance })
-      return
-    }
-
-    if (selection.mode === 'sketch_edit' && !pendingAdd && pendingSketchFilletRef.current && sketchEditPreviewRef.current) {
-      const featureId = selection.selectedFeatureId
-      const feature = featureId ? project.features.find((f) => f.id === featureId) ?? null : null
-      if (!feature) return
-      const radius = selection.sketchEditTool === 'chamfer'
-        ? chamferDistanceFromPoint(feature, pendingSketchFilletRef.current.anchorIndex, sketchEditPreviewRef.current.point)
-        : filletRadiusFromPoint(feature, pendingSketchFilletRef.current.anchorIndex, sketchEditPreviewRef.current.point)
-      fillet.setFilletDimensionEdit({ anchorIndex: pendingSketchFilletRef.current.anchorIndex, corner: pendingSketchFilletRef.current.corner, radius: radius ? formatLength(radius, units) : '' })
-      return
-    }
-
-    if (selection.mode === 'sketch_edit' && !pendingAdd && !fillet.filletDimensionEditRef.current) {
-      const currentEdit = dimEdit.dimensionEditRef.current
-      if (!currentEdit && dimEdit.dimensionEditControlRef.current) {
-        dimEdit.advanceTabInEditMode()
-      }
-    }
+    triggerDimensionEditFn({
+      project: projectRef.current,
+      pendingAdd: pendingAddRef.current,
+      pendingMove: pendingMoveRef.current,
+      pendingTransform: pendingTransformRef.current,
+      pendingOffset: pendingOffsetRef.current,
+      selectionMode: selectionRef.current.mode,
+      selectedFeatureId: selectionRef.current.selectedFeatureId,
+      sketchEditTool: selectionRef.current.sketchEditTool,
+      pendingPreviewPoint: pendingPreviewPointRef.current,
+      pendingMovePreviewPoint: pendingMovePreviewPointRef.current,
+      pendingTransformPreviewPoint: pendingTransformPreviewPointRef.current,
+      pendingOffsetRawPreviewPoint: pendingOffsetRawPreviewPointRef.current,
+      pendingOffsetPreviewPoint: pendingOffsetPreviewPointRef.current,
+      sketchEditPreviewPoint: sketchEditPreviewRef.current,
+      pendingSketchFillet: pendingSketchFilletRef.current,
+      units: projectRef.current.meta.units,
+      canvasWidth: canvasRef.current?.width ?? 0,
+      canvasHeight: canvasRef.current?.height ?? 0,
+      viewState: viewStateRef.current,
+      activeSnapMode: snap.activeSnapRef.current?.mode ?? null,
+      dimEdit,
+      move,
+      setOperationDimEdit,
+      fillet,
+    })
   }
 
   const keyboard = useCanvasKeyboard({
