+++ src/store/featureLifecycle.test.ts
@@ -35,6 +35,7 @@ import {
   type TextFeatureData,
 } from '../types/project'
 import { useProjectStore } from './projectStore'
+import type { ProjectStore } from './types'
 import { getDefinitionId } from './helpers/featureDefinitions'
 
 // ── Assertion helpers ──────────────────────────────────────────────
@@ -62,7 +63,7 @@ function resetStore(project?: Project): void {
     pendingConstraint: null,
     pendingTransform: null,
     pendingOffset: null,
-  } as any)
+  } as unknown as Partial<ProjectStore>)
 }
 
 function getProject(): Project {
